noticias_chile_bot/requirements.txt
feedparser
requests
tweepy
gspread
oauth2client
transformers
torch
python-dotenv
schedule
pillow